+++ changelog.md
@@ -1,1 +1,1 @@
-- Fixed converter sometimes not consuming potion
+- Replaced reflection with access transformers